Added commons-logging.properties in src/test/resources to force the JUnit tests to use the JDKLogger rather than the default Log4J logger.
Modified SurefireBooter so that it created the classpath arguments for the forked Surefire process as individual String array elements {"-classpath", "stuff in class path - can contain whitespace"} Configured in this manner, elements in the actual classpath may contain whitespace and forked processes will be executable.

Modified BooterTestCase and removed failure message if maven.repo.local contained embedded spaces.

Modified ForkedSurefireLoader and removed a manipulation of trhe classpath. The inclusion of classpath information within the class was a programmer error. A proper classpath must be used to fork the Surefire process, but the classpath information is not explivitly needed by the forked program, so it can be dropped.

git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/branches/SUREFIRE_FORK_EXEC/surefire@331390 13f79535-47bb-0310-9956-ffa450edef68