PR: MNG-1303 Submitted By: John Casey
Commented out the loadClass(..) method in IsolatedClassLoader, and added a new constructor that allows the user to define a parent classloader. Then, the surefire booter includes ClassLoader.getSystemClassLoader() as the parent to the isolated classloader it creates for testing...

NOTE: This may break some things, since the loadClass(..) method in IsolatedClassLoader reversed the normal parent-first loading strategy. I'm committing it so Jason can take a look, since he wrote the IsolatedClassLoader (or at least committed it).



git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/trunk/surefire-booter@349813 13f79535-47bb-0310-9956-ffa450edef68