PR: SUREFIRE-54
Reverted the revert of my previous fix, and fixed classloading for
TestNG. 

The surefireClassloader is now a child of the testsClassloader, meaning
that the tests also contain the testing framework, and surefire has access
to that.

In the case of TestNG, the annotations used to be in the two classloaders,
making them incompatible. Now they're only in the testsClassloader
and the surefireClassloader can safely instantiate and run TestNG tests.

Also childDelegation is deprecated now - it is no longer needed since
the testsClassloaders have no parent. Code is still here, just in case.


git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/trunk@441092 13f79535-47bb-0310-9956-ffa450edef68