[SUREFIRE-584] Fix the verifier based tests so that they work during mvn release:prepare
o I have used invoker:install to stage the artifacts required for the integration tests

o I use a JUnit test to setup the settings.xml that uses the staged local repository and uses the actual local repository as a remote repository while retaining all the running machine's settings (e.g. proxies) that should mean that the integration tests will work from behind a corporate proxy and they should be fast

o I use both surefire and failsafe maven plugins for test execution, failsafe is used for the integration tests and surefire runs the JUnit test that creates the settings.xml for the staged local repository.  I could have just used exec-maven-plugin and used a executible java class to create the staged settings.xml, but my concern is that that would introduce another dependency and complicate bootstrapping... failsafe will not be an issue in this respect as I will be adding it to surefire.

o All the integration tests have been renamed to be picked up by failsafe's default patterns.

git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/trunk@896852 13f79535-47bb-0310-9956-ffa450edef68