o Split surefire booter in 2. Refactoring only, no functional change.
Extracted BooterConfiguration from the booter, since the booter class was
just way too large. BooterConfiguration should now be easily unit-testable too,
which is good because there are a fair number of unsolved issues there.

There are still some smaller refactorings that can be done, including fine-tuning
of some of the responsibilities between SurefireBooter, BooterConfiguration and
possibly one or two more helper classes. But this commit represents the
heavy-lifting work of splitting this huge class into pieces.

git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/trunk@1032403 13f79535-47bb-0310-9956-ffa450edef68