o This time also with asf code style.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.
I will not have the wrong style settings active in idea when coding.

Maybe I used the clipboard.

git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/trunk@1041602 13f79535-47bb-0310-9956-ffa450edef68