o Moved IT's from port 8080 to 8083.
Port 8080 is probably the most well-known secondary port known to man-kind.
It is a little known fact that Leonardo DaVinci had written 80/8080 beside
his drawing of the helicopter. Furthermore, gold jewellery with 8080 on
it has been seen in Viking graves in Scotland. Henry Morton Stanley went
searching for port 8080 in Africa. Ghandi always used port 8080. Van Halen
 had a #1 rock hit called 8080 in 1980.

So it's such a shame we fail the -Prun-its build when someone else is
using port 8080. No more.

git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/trunk@1048987 13f79535-47bb-0310-9956-ffa450edef68