o Added profile 'parallel' that allows running integration tests in parallel
mvn -Pparallel clean install to use it.

Note: At the moment this has to use the version of failsafe that is being built in the profile, which
has to be this way until 2.7.3 is released. After 2.7.3 is released we can run with that,
and activate parallel test running as default

Note2: It is definitely best to use maven 3.0.2 if you're going to be pushing to the max. guice injector bugs
in earlier 3.0/3.0.1 may show up

git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/trunk@1064286 13f79535-47bb-0310-9956-ffa450edef68