o Changed build to use shaded provider itroduced in SUREFIRE-700.
From now on we have isolation. Note that I had to change
the integration tests to run with the *current* version
of surefire. This is because the integration test module
uses the 'install' feature of the maven-invoker-plugin,
which requires that the necessary artifacts are on the project
classpath, including the providers. Unfortunately, the
shaded provider does not like the presence of the other
providers. This problem can/should be fixed by extracting
the maven-invoker-plugin into a separate module, so we can
avoid the weird dependencies creeping into the IT-runner.

Given the level of weirdness any surefire developer
has had to cope with, this should be peanuts. Will
fix real soon, anyway ;)

git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/trunk@1081113 13f79535-47bb-0310-9956-ffa450edef68