[SUREFIRE-933] Tests not matching categories would fail build
Fixed in a way that does not break forkMode=onceperthread (enhanced the IT for that), refined the logic of JUnitCoreProvider.canRunClass (to also detect the BasicTest class from the IT)
Tests are now filtered right before handing them over to JUnit. In case of a NoTestsRemainException while filtering the test request, nothing will be handed over to JUnit.