SUREFIRE-569 Add support for scanning Dependencies for TestClasses
Configure Surefire to scan any project Dependency for TestClasses
to include in the given test run.

TestClasses found as part of a Dependency are handled in the same
way as the testClassesDirectory found test classes. Includes,
Excludes and Test filters work the same.

Configuration example:

<dependency>
  <groupId>org.a</groupId>
  <artifactId>a</artifactId>
</dependency>

<configuration>
  <dependenciesToScan>
    <dependency>org.a:a</dependency>
  </dependenciesToScan>
</configuration>

Patch applied with rename of 1 file.