Add @{...} property evaluation for the argLine.
Replaces expressions @{property-name} in the argLine with the
corresponding properties from the model. This allows late evaluation
of property values when the plugin is executed (as compared to
evaluation when the pom is parsed as is done with ${property-name}
expressions).

This allows other plugins to modify or set properties with the changes
getting picked up by surefire.