Fixing build on Windows cmd.exe.
Files generated with the XJC tool were using the platform encoding, that can be different from the project source encoding. In this case, the project source encoding is UTF-8, but cmd.exe can use another (e.g. code page 850). In turn, this created an error while compiling the generated sources because they contain unmappable characters for UTF-8 as expected by the compiler. The encoding that should be used by XJC is specified through the -encoding option.