Support for externally passed random seed
It's good practice to execute tests in random order. That ensures tests are
atomic and not depend on each other or some stateful entity. But, if such error
occurs, user should be able to reproduce this exact, erroneous execution.

This commit adds the ability to reproduce those errors by adding support for
externally passed random seed. If random seed is not passed, it will be
generated automatically. Also, random seed will be printed on console. In that
way used seed can be achieved by CI server for later.

This operation is strongly influenced by Ruby's rspec order command line option:

https://www.relishapp.com/rspec/rspec-core/v/2-13/docs/command-line/order-new-in-rspec-core-2-8

Showcase of execution:

https://asciinema.org/a/42342

Remarks:

 * Fixing flaky test Surefire1177TestngParallelSuitesIT

   This test rely on displaying messages on System.out, but invocations
   wasn't synchronized so messeges tend to interweave and test wasn't
   able to assert for them properly. I added synchronization and move
   code a litle bit.