[SUREFIRE-1840] Distinguish commands in docs
sudo, which provides root/ administrator level priviledges on *nix
systems, should be only used when required. Given that docker uses the
kernel of the host system for Linux while docker uses a virtual machine
for Windows/macOS, it is omitted for the latter. This was clarified for
the documentation concerning testing Maven Surefire with docker.