[SUREFIRE-2136] Rename test dir that has colon
In the UmlautDirIT two special directories are created that have a
colon character in the name. This causes problems with the artifact
upload of the Github actions, so we rename those directories in a
finally block at the end of the test to something without a colon
character in it.