[SUREFIRE-2152] Include display-name of JUnit5 templated-tests in methodDesc
This includes the displayName (e.g. value of name="..." of a
@ParameterizedTest) in the method description / test identifier.

Provides more information than merely the invocation ID in case
of failing tests, also for the console reporter. This can
(especially if the actual list of arguments are quite large and/or
dynamically determined at runtime) help a lot to identify which
actual cases are the problematic ones.