Create a single LauncherSession for invocations of JUnitPlatformProvider (#863)
Prior to this commit, `JUnitPlatformProvider` created multiple
`LauncherSessions` for some executions, for example when `forkTestSet`
was `null` or when rerunning failed tests. This caused registered
`LauncherSessionListeners` registered via `ServiceLoader` on the JUnit
Platform to be called multiple times. As they are intended to set up and
tear down resources, this had the potential of unnecessarily slowing
down test execution.